Move pkg/lib/probe to pkg/lib
